// src/lib/sanity/types.ts
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};

export type AccessibleImage = {
  _type: "accessibleImage";
  asset?: SanityImageAssetReference;
  media?: unknown;
  hotspot?: SanityImageHotspot;
  crop?: SanityImageCrop;
  alt?: string;
  caption?: string;
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type AMENITIES_QUERY_RESULT = Array<{
  _id: string;
  name: string | null;
  category: string | null;
  description: string | null;
  image: AccessibleImage | null;
  icon:
    | "accessibility"
    | "bed"
    | "dining"
    | "leaf"
    | "music"
    | "parking"
    | "pet"
    | "water"
    | null;
}>;